applied a patch from devmail@hispeed.ch:
---
Hi Kohsuke,

Akuma, your "Java process daemonizer" is good, but it annoyed me that there 
was no way for me to decide the name and location of the PID file.

I have attached a patch which implements a straightforward way of achieving 
this functionality, while still preserving the old API and semantics.

Feel free to apply the patch as you see fit. I hope you find it useful, and 
that functionality like this finds its way into the next version.

Regards,
Olav Reinert



git-svn-id: https://akuma.dev.java.net/svn/akuma/trunk/akuma@47 f4debed6-9561-0410-b0a4-b640bb5a59ad